+++ src/main/java/ro/danix/first/model/repository/BabyRepository.java
@@ -1,6 +1,6 @@
 package ro.danix.first.model.repository;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import ro.danix.first.model.domain.Baby;
@@ -10,7 +10,7 @@ import ro.danix.first.model.domain.user.BabyParent;
  *
  * @author danix
  */
-public interface BabyRepository extends MongoRepository<Baby, BigDecimal> {
+public interface BabyRepository extends MongoRepository<Baby, BigInteger> {
 
     List<Baby> findByBabyParent(BabyParent babyParent);
 }
