+++ src/main/java/ro/danix/first/model/repository/user/BabyParentRepository.java
@@ -1,6 +1,6 @@
 package ro.danix.first.model.repository.user;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import ro.danix.first.model.domain.EmailAddress;
 import ro.danix.first.model.domain.user.BabyParent;
@@ -9,7 +9,7 @@ import ro.danix.first.model.domain.user.BabyParent;
  *
  * @author danix
  */
-public interface BabyParentRepository extends MongoRepository<BabyParent, BigDecimal> {
+public interface BabyParentRepository extends MongoRepository<BabyParent, BigInteger> {
 
     BabyParent findByEmailAddress(EmailAddress emailAddress);
 }
