+++ src/main/java/ro/danix/first/model/repository/user/UserRepository.java
@@ -1,6 +1,6 @@
 package ro.danix.first.model.repository.user;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import ro.danix.first.model.domain.EmailAddress;
 import ro.danix.first.model.domain.user.User;
@@ -9,7 +9,7 @@ import ro.danix.first.model.domain.user.User;
  *
  * @author danix
  */
-public interface UserRepository extends MongoRepository<User, BigDecimal> {
+public interface UserRepository extends MongoRepository<User, BigInteger> {
 
     User findByEmailAddress(EmailAddress emailAddress);
 }
