+++ src/main/java/ro/danix/first/model/service/BabyService.java
@@ -1,6 +1,6 @@
 package ro.danix.first.model.service;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import ro.danix.first.model.domain.Baby;
 
@@ -8,5 +8,5 @@ import ro.danix.first.model.domain.Baby;
  *
  * @author danix
  */
-public interface BabyService extends MongoRepository<Baby, BigDecimal> {
+public interface BabyService extends MongoRepository<Baby, BigInteger> {
 }
