+++ src/main/java/ro/danix/first/model/service/user/BabyParentService.java
@@ -1,6 +1,6 @@
 package ro.danix.first.model.service.user;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import ro.danix.first.model.domain.user.BabyParent;
 
@@ -8,7 +8,7 @@ import ro.danix.first.model.domain.user.BabyParent;
  *
  * @author danix
  */
-public interface BabyParentService extends MongoRepository<BabyParent, BigDecimal> {
+public interface BabyParentService extends MongoRepository<BabyParent, BigInteger> {
 
     void follow(BabyParent babyParent, BabyParent follower);
 }
