+++ src/main/java/ro/danix/first/model/service/user/UserService.java
@@ -1,6 +1,6 @@
 package ro.danix.first.model.service.user;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import ro.danix.first.model.domain.user.User;
 
@@ -8,7 +8,7 @@ import ro.danix.first.model.domain.user.User;
  *
  * @author danix
  */
-public interface UserService extends MongoRepository<User, BigDecimal> {
+public interface UserService extends MongoRepository<User, BigInteger> {
 
     void follow(User user, User follower);
 }
